moved callbacks to props obj and added invalidname
