+++ src/components/forms/cbr/CbrResourceAttributeForm.js
@@ -29,9 +29,8 @@ class CbrResourceAttributeForm extends Component {
           componentName={this.props.data.name + "-cbr-ra"}
           value={this.state.name}
           onChange={this.handleInputChange}
-          invalid={this.props.invalidCallback("name", this.state, this.props)}
-          invalidText={this.props.invalidTextCallback(
-            "name",
+          invalid={this.props.invalidNameCallback(this.state, this.props)}
+          invalidText={this.props.invalidNameTextCallback(
             this.state,
             this.props
           )}
@@ -66,11 +65,12 @@ CbrResourceAttributeForm.defaultProps = {
 };
 
 CbrResourceAttributeForm.propTypes = {
-  isModal: PropTypes.bool,
   data: PropTypes.shape({
     name: PropTypes.string.isRequired,
     value: PropTypes.string.isRequired,
   }),
+  invalidNameCallback: PropTypes.func.isRequired,
+  invalidNameTextCallback: PropTypes.func.isRequired,
   invalidCallback: PropTypes.func.isRequired,
   invalidTextCallback: PropTypes.func.isRequired,
   arrayParentName: PropTypes.string,
