+++ src/components/forms/cbr/CbrRuleForm.js
@@ -34,8 +34,6 @@ class CbrRuleForm extends Component {
   render() {
     // set up props for subforms
     let contextInnerFormProps = {
-      invalidCallback: this.props.invalidContextCallback,
-      invalidTextCallback: this.props.invalidContextTextCallback,
       arrayParentName: this.props.data.name,
     };
     transpose({ ...this.props.contextProps }, contextInnerFormProps);
@@ -66,9 +64,8 @@ class CbrRuleForm extends Component {
             value={this.state.name}
             onChange={this.handleInputChange}
             hideHelperText={true}
-            invalid={this.props.invalidCallback("name", this.state, this.props)}
-            invalidText={this.props.invalidTextCallback(
-              "name",
+            invalid={this.props.invalidNameCallback(this.state, this.props)}
+            invalidText={this.props.invalidNameTextCallback(
               this.state,
               this.props
             )}
@@ -143,6 +140,7 @@ class CbrRuleForm extends Component {
               tooltip={{
                 content: "Contexts define where your resource can be accessed.",
                 link: "https://cloud.ibm.com/docs/account?topic=account-context-restrictions-whatis#restriction-context",
+                align: "top-left",
               }}
               addText="Create a Context"
               arrayData={this.props.data.contexts}
@@ -192,7 +190,8 @@ class CbrRuleForm extends Component {
               name="Tags"
               subHeading
               tooltip={{
-                content: "A cloud resource can also be specified using tags",
+                content: "A cloud resource can also be specified using tags.",
+                align: "top-left",
               }}
               addText="Create a Tag"
               arrayData={this.props.data.tags}
@@ -222,7 +221,7 @@ CbrRuleForm.defaultProps = {
   data: {
     name: "",
     description: "",
-    enforcement_mode: "",
+    enforcement_mode: "Enabled",
     api_type_id: "",
     contexts: [],
     resource_attributes: [],
@@ -257,33 +256,44 @@ CbrRuleForm.propTypes = {
       }).isRequired
     ),
   }),
-  invalidTextCallback: PropTypes.func.isRequired,
+  propsMatchState: PropTypes.func.isRequired,
   invalidCallback: PropTypes.func.isRequired,
+  invalidTextCallback: PropTypes.func.isRequired,
+  invalidNameCallback: PropTypes.func.isRequired,
+  invalidNameTextCallback: PropTypes.func.isRequired,
   contextProps: PropTypes.shape({
-    isModal: PropTypes.bool,
-    data: PropTypes.shape({
-      name: PropTypes.string.isRequired,
-      value: PropTypes.string.isRequired,
-    }),
+    onSave: PropTypes.func.isRequired,
+    onDelete: PropTypes.func.isRequired,
+    onSubmit: PropTypes.func.isRequired,
+    disableSave: PropTypes.func.isRequired,
+    invalidCallback: PropTypes.func.isRequired,
+    invalidTextCallback: PropTypes.func.isRequired,
+    invalidNameCallback: PropTypes.func.isRequired,
+    invalidNameTextCallback: PropTypes.func.isRequired,
   }).isRequired,
   invalidContextCallback: PropTypes.func.isRequired,
   invalidContextTextCallback: PropTypes.func.isRequired,
   resourceAttributeProps: PropTypes.shape({
-    isModal: PropTypes.bool,
-    data: PropTypes.shape({
-      name: PropTypes.string.isRequired,
-      value: PropTypes.string.isRequired,
-    }),
+    onSave: PropTypes.func.isRequired,
+    onDelete: PropTypes.func.isRequired,
+    onSubmit: PropTypes.func.isRequired,
+    disableSave: PropTypes.func.isRequired,
+    invalidCallback: PropTypes.func.isRequired,
+    invalidTextCallback: PropTypes.func.isRequired,
+    invalidNameCallback: PropTypes.func.isRequired,
+    invalidNameTextCallback: PropTypes.func.isRequired,
   }).isRequired,
   invalidResourceAttributeCallback: PropTypes.func.isRequired,
   invalidResourceAttributeTextCallback: PropTypes.func.isRequired,
   tagProps: PropTypes.shape({
-    isModal: PropTypes.bool,
-    data: PropTypes.shape({
-      name: PropTypes.string.isRequired,
-      operator: PropTypes.string.isRequired,
-      value: PropTypes.string.isRequired,
-    }),
+    onSave: PropTypes.func.isRequired,
+    onDelete: PropTypes.func.isRequired,
+    onSubmit: PropTypes.func.isRequired,
+    disableSave: PropTypes.func.isRequired,
+    invalidCallback: PropTypes.func.isRequired,
+    invalidTextCallback: PropTypes.func.isRequired,
+    invalidNameCallback: PropTypes.func.isRequired,
+    invalidNameTextCallback: PropTypes.func.isRequired,
   }).isRequired,
   invalidTagCallback: PropTypes.func.isRequired,
   invalidTagTextCallback: PropTypes.func.isRequired,
