+++ src/components/forms/cbr/CbrTagForm.js
@@ -33,9 +33,8 @@ class CbrTagForm extends Component {
           className={"fieldWidthSmaller"}
           value={this.state.name}
           onChange={this.handleInputChange}
-          invalid={this.props.invalidCallback("name", this.state, this.props)}
-          invalidText={this.props.invalidTextCallback(
-            "name",
+          invalid={this.props.invalidNameCallback(this.state, this.props)}
+          invalidText={this.props.invalidNameTextCallback(
             this.state,
             this.props
           )}
@@ -92,12 +91,13 @@ CbrTagForm.defaultProps = {
 };
 
 CbrTagForm.propTypes = {
-  isModal: PropTypes.bool,
   data: PropTypes.shape({
     name: PropTypes.string.isRequired,
     operator: PropTypes.string.isRequired,
     value: PropTypes.string.isRequired,
   }),
+  invalidNameCallback: PropTypes.func.isRequired,
+  invalidNameTextCallback: PropTypes.func.isRequired,
   invalidCallback: PropTypes.func.isRequired,
   invalidTextCallback: PropTypes.func.isRequired,
   arrayParentName: PropTypes.string,
