+++ storybook/src/stories/Forms/ContextBasedRestrictions/CbrResourceAttributeForm.stories.js
@@ -1,40 +1,40 @@
 import React from "react";
-import { CbrContextForm } from "icse-react-assets";
+import { CbrResourceAttributeForm } from "icse-react-assets";
 import { contains } from "lazy-z";
 
 export default {
-  component: CbrContextForm,
-  title: "Components/Forms/CBR/CbrContextForm",
+  component: CbrResourceAttributeForm,
+  title: "Components/Forms/ContextBasedRestrictions/CbrResourceAttributeForm",
   args: {
     isModal: false,
   },
   argTypes: {
     isModal: {
       description:
-        "A boolean value that specifies if the CbrContextForm is a component within a modal",
+        "A boolean value that specifies if the CbrResourceAttributeForm is a component within a modal",
       type: { required: false },
       control: "boolean",
       table: { defaultValue: { summary: "false" } },
     },
     "data.name": {
-      description: "String, the name of the context",
+      description: "String, the name of the resource attribute",
       control: "none",
       type: { required: true },
     },
     "data.value": {
-      description: "String, the value for the context",
+      description: "String, the value for the resource attribute",
       control: "none",
       type: { required: true },
     },
     invalidCallback: {
       description:
-        "Function that determines invalid state for the context's name and value, returns a boolean",
+        "Function that determines invalid state for the attribute's name and value, returns a boolean",
       type: { required: true },
       control: "none",
     },
     invalidTextCallback: {
       description:
-        "Function that returns invalid text string for the context's name and value",
+        "Function that returns invalid text string for the resource attribute's name and value",
       type: { required: true },
       control: "none",
     },
@@ -42,7 +42,7 @@ export default {
   parameters: {
     docs: {
       description: {
-        component: "CbrContextForm",
+        component: "CbrResourceAttributeForm",
       },
     },
     source: {
@@ -51,7 +51,7 @@ export default {
   },
 };
 
-const CbrContextFormStory = ({ ...args }) => {
+const CbrResourceAttributeFormStory = ({ ...args }) => {
   function validName(str) {
     const regex = /^[A-z]([a-z0-9-]*[a-z0-9])?$/i;
     if (str) return str.match(regex) !== null;
@@ -74,10 +74,10 @@ const CbrContextFormStory = ({ ...args }) => {
     else return `Invalid ${field}`;
   }
   return (
-    <CbrContextForm
+    <CbrResourceAttributeForm
       {...args}
       data={{
-        name: "context",
+        name: "resource-attribute",
         value: "example",
       }}
       invalidCallback={invalidCallback}
@@ -86,4 +86,4 @@ const CbrContextFormStory = ({ ...args }) => {
   );
 };
 
-export const Default = CbrContextFormStory.bind({});
+export const Default = CbrResourceAttributeFormStory.bind({});
